feat: export getValueColor from chartUtils
