+++ client/dashboard/src/components/chart/MetricCard.tsx
@@ -1,4 +1,5 @@
 import { cn, Icon, type IconName } from "@speakeasy-api/moonshine";
+import { getValueColor, ThresholdConfig } from "./chartUtils";
 
 type AccentColor = "red" | "orange" | "yellow" | "green" | "blue" | "purple";
 
@@ -14,31 +15,6 @@ export type MetricCardProps = {
   accentColor?: AccentColor;
   subtext?: string;
 };
-
-export type ThresholdConfig = {
-  red: number;
-  amber: number;
-  inverted?: boolean; // true if lower is better (like latency)
-};
-
-export function getValueColor(
-  value: number,
-  thresholds?: ThresholdConfig,
-): string {
-  if (!thresholds) return "";
-
-  if (thresholds.inverted) {
-    // Lower is better (e.g., latency)
-    if (value > thresholds.red) return "text-red-500";
-    if (value > thresholds.amber) return "text-amber-500";
-    return "text-emerald-600";
-  } else {
-    // Higher is better (e.g., chats, resolution rate)
-    if (value < thresholds.red) return "text-red-500";
-    if (value < thresholds.amber) return "text-amber-500";
-    return "text-emerald-600";
-  }
-}
 
 const accentColorsMap: Record<AccentColor, string> = {
   red: "border-t-red-500",
