+++ client/dashboard/src/pages/observability/ObservabilityOverview.tsx
@@ -1420,28 +1420,6 @@ function ObservabilityContent({
   );
 }
 
-type ThresholdConfig = {
-  red: number;
-  amber: number;
-  inverted?: boolean; // true if lower is better (like latency)
-};
-
-function getValueColor(value: number, thresholds?: ThresholdConfig): string {
-  if (!thresholds) return "";
-
-  if (thresholds.inverted) {
-    // Lower is better (e.g., latency)
-    if (value > thresholds.red) return "text-red-500";
-    if (value > thresholds.amber) return "text-amber-500";
-    return "text-emerald-600";
-  } else {
-    // Higher is better (e.g., chats, resolution rate)
-    if (value < thresholds.red) return "text-red-500";
-    if (value < thresholds.amber) return "text-amber-500";
-    return "text-emerald-600";
-  }
-}
-
 // Chart selection wrapper for drag-to-zoom functionality
 function ChartWithSelection({
   children,
